--HG--
rename : browser/base/content/tab-content.js => browser/actors/AboutReaderChild.jsm
rename : browser/modules/BlockedSiteContent.jsm => browser/actors/BlockedSiteChild.jsm
rename : browser/modules/ClickEventHandler.jsm => browser/actors/ClickHandlerChild.jsm
rename : browser/base/content/tab-content.js => browser/actors/ContentSearchChild.jsm
rename : browser/modules/ContextMenu.jsm => browser/actors/ContextMenuChild.jsm
rename : browser/base/content/tab-content.js => browser/actors/DOMFullscreenChild.jsm
rename : browser/modules/LightWeightThemeWebInstallListener.jsm => browser/actors/LightWeightThemeInstallChild.jsm
rename : browser/modules/NetErrorContent.jsm => browser/actors/NetErrorChild.jsm
rename : browser/base/content/content.js => browser/actors/OfflineAppsChild.jsm
rename : browser/modules/PageInfoListener.jsm => browser/actors/PageInfoChild.jsm
rename : browser/base/content/content.js => browser/actors/PageMetadataChild.jsm
rename : browser/modules/PageStyleHandler.jsm => browser/actors/PageStyleChild.jsm
rename : browser/modules/ContentWebRTC.jsm => browser/actors/WebRTCChild.jsm
rename : browser/components/uitour/ContentUITour.jsm => browser/components/uitour/UITourChild.jsm
rename : browser/extensions/formautofill/bootstrap.js => browser/extensions/formautofill/api.js
rename : dom/base/ProcessGlobal.cpp => dom/base/ContentProcessMessageManager.cpp
rename : dom/base/ProcessGlobal.h => dom/base/ContentProcessMessageManager.h
rename : dom/base/nsInProcessTabChildGlobal.cpp => dom/base/InProcessTabChildMessageManager.cpp
rename : dom/base/nsInProcessTabChildGlobal.h => dom/base/InProcessTabChildMessageManager.h
rename : dom/ipc/ManifestMessages.jsm => dom/ipc/ManifestMessagesChild.jsm
rename : toolkit/content/browser-content.js => toolkit/actors/AudioPlaybackChild.jsm
rename : toolkit/modules/DateTimePickerContent.jsm => toolkit/actors/DateTimePickerChild.jsm
rename : toolkit/content/browser-content.js => toolkit/actors/FindBarChild.jsm
rename : toolkit/modules/RemoteFinder.jsm => toolkit/actors/FinderChild.jsm
rename : toolkit/modules/PopupBlocking.jsm => toolkit/actors/PopupBlockingChild.jsm
rename : toolkit/modules/PrintingContent.jsm => toolkit/actors/PrintingChild.jsm
rename : toolkit/modules/SelectionSourceContent.jsm => toolkit/actors/SelectionSourceChild.jsm
rename : toolkit/modules/WebChannelContent.jsm => toolkit/actors/WebChannelChild.jsm
rename : toolkit/modules/WebNavigationChild.jsm => toolkit/actors/WebNavigationChild.jsm
rename : toolkit/components/normandy/content/ShieldFrameListener.jsm => toolkit/components/normandy/content/ShieldFrameChild.jsm
rename : toolkit/content/tests/widgets/mochitest.ini => toolkit/content/tests/widgets/xbl/mochitest.ini
rename : toolkit/content/tests/widgets/test_videocontrols.html => toolkit/content/tests/widgets/xbl/test_videocontrols.html
rename : toolkit/content/tests/widgets/test_videocontrols_error.html => toolkit/content/tests/widgets/xbl/test_videocontrols_error.html
rename : toolkit/content/tests/widgets/test_videocontrols_size.html => toolkit/content/tests/widgets/xbl/test_videocontrols_size.html
rename : toolkit/content/tests/widgets/test_videocontrols_vtt.html => toolkit/content/tests/widgets/xbl/test_videocontrols_vtt.html
rename : toolkit/content/widgets/videocontrols.xml => toolkit/content/widgets/videocontrols.js
rename : toolkit/modules/FindBarChild.jsm => toolkit/modules/FindBarContent.jsm
rename : toolkit/modules/RemoteFinder.jsm => toolkit/modules/FinderParent.jsm

+++ mobile/android/base/java/org/mozilla/gecko/GeckoApp.java
@@ -7,6 +7,7 @@ package org.mozilla.gecko;
 
 import org.mozilla.gecko.AppConstants.Versions;
 import org.mozilla.gecko.GeckoProfileDirectories.NoMozillaDirectoryException;
+import org.mozilla.gecko.GeckoScreenOrientation.ScreenOrientation;
 import org.mozilla.gecko.annotation.RobocopTarget;
 import org.mozilla.gecko.annotation.WrapForJNI;
 import org.mozilla.gecko.health.HealthRecorder;
@@ -40,7 +41,6 @@ import org.mozilla.gecko.util.ThreadUtils;
 import org.mozilla.gecko.util.ViewUtil;
 import org.mozilla.gecko.widget.ActionModePresenter;
 import org.mozilla.gecko.widget.AnchoredPopup;
-import org.mozilla.geckoview.GeckoRuntime;
 import org.mozilla.geckoview.GeckoSession;
 import org.mozilla.geckoview.GeckoSessionSettings;
 import org.mozilla.geckoview.GeckoView;
@@ -54,6 +54,7 @@ import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
+import android.content.pm.ApplicationInfo;
 import android.content.pm.PackageManager.NameNotFoundException;
 import android.content.res.Configuration;
 import android.graphics.BitmapFactory;
@@ -113,6 +114,7 @@ public abstract class GeckoApp extends GeckoActivity
                                           BundleEventListener,
                                           GeckoMenu.Callback,
                                           GeckoMenu.MenuPresenter,
+                                          GeckoScreenOrientation.OrientationChangeListener,
                                           GeckoSession.ContentDelegate,
                                           ScreenOrientationDelegate,
                                           Tabs.OnTabsChangedListener,
@@ -982,6 +984,7 @@ public abstract class GeckoApp extends GeckoActivity
         } catch (ClassNotFoundException e) { }
 
         GeckoAppShell.setScreenOrientationDelegate(this);
+        GeckoScreenOrientation.getInstance().addListener(this);
 
         // Tell Stumbler to register a local broadcast listener to listen for preference intents.
         // We do this via intents since we can't easily access Stumbler directly,
@@ -1064,8 +1067,6 @@ public abstract class GeckoApp extends GeckoActivity
         }
 
         super.onCreate(savedInstanceState);
-
-        GeckoScreenOrientation.getInstance().update(getResources().getConfiguration().orientation);
 
         setContentView(getLayout());
 
@@ -1930,11 +1931,6 @@ public abstract class GeckoApp extends GeckoActivity
 
         GeckoAppShell.setScreenOrientationDelegate(this);
 
-        int newOrientation = getResources().getConfiguration().orientation;
-        if (GeckoScreenOrientation.getInstance().update(newOrientation)) {
-            refreshChrome();
-        }
-
         // We use two times: a pseudo-unique wall-clock time to identify the
         // current session across power cycles, and the elapsed realtime to
         // track the duration of the session.
@@ -2134,6 +2130,8 @@ public abstract class GeckoApp extends GeckoActivity
         Tabs.unregisterOnTabsChangedListener(this);
         Tabs.getInstance().detachFromContext();
 
+        GeckoScreenOrientation.getInstance().removeListener(this);
+
         if (mShutdownOnDestroy) {
             GeckoApplication.shutdown(!mRestartOnShutdown ? null : new Intent(
                     Intent.ACTION_MAIN, /* uri */ null, getApplicationContext(), getClass()));
@@ -2176,19 +2174,6 @@ public abstract class GeckoApp extends GeckoActivity
         final Locale changed = localeManager.onSystemConfigurationChanged(this, getResources(), newConfig, mLastLocale);
         if (changed != null) {
             onLocaleChanged(Locales.getLanguageTag(changed));
-        }
-
-        // onConfigurationChanged is not called for 180 degree orientation changes,
-        // we will miss such rotations and the screen orientation will not be
-        // updated.
-        if (GeckoScreenOrientation.getInstance().update(newConfig.orientation)) {
-            if (mFormAssistPopup != null)
-                mFormAssistPopup.hide();
-            refreshChrome();
-        }
-
-        if (mPromptService != null) {
-            mPromptService.changePromptOrientation(newConfig.orientation);
         }
 
         super.onConfigurationChanged(newConfig);
@@ -2256,8 +2241,9 @@ public abstract class GeckoApp extends GeckoActivity
                 // the res/fonts directory: we no longer need to copy our
                 // bundled fonts out of the APK in order to use them.
                 // See https://bugzilla.mozilla.org/show_bug.cgi?id=878674.
-                File dir = new File("res/fonts");
-                if (dir.exists() && dir.isDirectory()) {
+                final File dataDir = new File(context.getApplicationInfo().dataDir);
+                final File dir = new File(dataDir, "res/fonts");
+                if (dir.exists() && dir.isDirectory() && dir.listFiles() != null) {
                     for (File file : dir.listFiles()) {
                         if (file.isFile() && file.getName().endsWith(".ttf")) {
                             file.delete();
@@ -2527,5 +2513,13 @@ public abstract class GeckoApp extends GeckoActivity
         }
         setRequestedOrientation(requestedActivityInfoOrientation);
         return true;
+    }
+
+    @Override
+    public void onScreenOrientationChanged(ScreenOrientation newOrientation) {
+        if (mFormAssistPopup != null) {
+            mFormAssistPopup.hide();
+        }
+        refreshChrome();
     }
 }
